refactor(Channels Providers JS): Moved all the channels an providers to their own directories for easier reading
