+++ ui/src/main/js/MainPage.js
@@ -7,26 +7,26 @@ import AboutInfo from 'component/AboutInfo';
 import LogoutConfirmation from 'component/common/LogoutConfirmation';
 import { getDescriptors } from 'store/actions/descriptors';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import SlackGlobalConfiguration from 'global/channels/slack/SlackGlobalConfiguration';
-import EmailGlobalConfiguration from 'global/channels/email/EmailGlobalConfiguration';
-import JiraCloudGlobalConfiguration from 'global/channels/jira/cloud/JiraCloudGlobalConfiguration';
-import { JIRA_CLOUD_INFO } from 'global/channels/jira/cloud/JiraCloudModel';
-import { SLACK_INFO } from 'global/channels/slack/SlackModels';
-import { EMAIL_INFO } from 'global/channels/email/EmailModels';
-import { JIRA_SERVER_INFO } from 'global/channels/jira/server/JiraServerModel';
-import { MSTEAMS_INFO } from 'global/channels/msteams/MSTeamsModel';
-import MSTeamsGlobalConfiguration from 'global/channels/msteams/MSTeamsGlobalConfiguration';
-import { AZURE_INFO } from 'global/channels/azure/AzureModel';
-import AzureGlobalConfiguration from 'global/channels/azure/AzureGlobalConfiguration';
+import SlackGlobalConfiguration from 'channels/slack/SlackGlobalConfiguration';
+import EmailGlobalConfiguration from 'channels/email/EmailGlobalConfiguration';
+import JiraCloudGlobalConfiguration from 'channels/jira/cloud/JiraCloudGlobalConfiguration';
+import { JIRA_CLOUD_INFO } from 'channels/jira/cloud/JiraCloudModel';
+import { SLACK_INFO } from 'channels/slack/SlackModels';
+import { EMAIL_INFO } from 'channels/email/EmailModels';
+import { JIRA_SERVER_INFO } from 'channels/jira/server/JiraServerModel';
+import { MSTEAMS_INFO } from 'channels/msteams/MSTeamsModel';
+import MSTeamsGlobalConfiguration from 'channels/msteams/MSTeamsGlobalConfiguration';
+import { AZURE_INFO } from 'channels/azure/AzureModel';
+import AzureGlobalConfiguration from 'channels/azure/AzureGlobalConfiguration';
 import { SCHEDULING_INFO } from 'global/components/scheduling/SchedulingModel';
 import SchedulingConfiguration from 'global/components/scheduling/SchedulingConfiguration';
 import { SETTINGS_INFO } from 'global/components/settings/SettingsModel';
 import SettingsConfiguration from 'global/components/settings/SettingsConfiguration';
 import { AUTHENTICATION_INFO } from 'global/components/auth/AuthenticationModel';
 import AuthenticationConfiguration from 'global/components/auth/AuthenticationConfiguration';
-import { BLACKDUCK_INFO, BLACKDUCK_URLS } from 'global/providers/blackduck/BlackDuckModel';
-import BlackDuckProviderConfiguration from 'global/providers/blackduck/BlackDuckProviderConfiguration';
-import BlackDuckConfiguration from 'global/providers/blackduck/BlackDuckConfiguration';
+import { BLACKDUCK_INFO, BLACKDUCK_URLS } from 'providers/blackduck/BlackDuckModel';
+import BlackDuckProviderConfiguration from 'providers/blackduck/BlackDuckProviderConfiguration';
+import BlackDuckConfiguration from 'providers/blackduck/BlackDuckConfiguration';
 import { AUDIT_INFO } from 'global/components/audit/AuditModel';
 import AuditPage from 'dynamic/loaded/audit/AuditPage';
 import { CERTIFICATE_INFO } from 'global/components/certificates/CertificateModel';
@@ -35,7 +35,7 @@ import { TASK_MANAGEMENT_INFO } from 'global/components/task/TaskManagementModel
 import TaskManagement from 'dynamic/loaded/tasks/TaskManagement';
 import { USER_MANAGEMENT_INFO } from 'global/components/user/UserModel';
 import UserManagement from 'dynamic/loaded/users/UserManagement';
-import JiraServerGlobalConfiguration from 'global/channels/jira/server/JiraServerGlobalConfiguration';
+import JiraServerGlobalConfiguration from 'channels/jira/server/JiraServerGlobalConfiguration';
 import { doesDescriptorExist } from 'util/descriptorUtilities';
 import { DISTRIBUTION_INFO, DISTRIBUTION_URLS } from 'distribution/DistributionModel';
 import DistributionConfiguration from 'distribution/DistributionConfiguration';
