+++ ui/src/main/js/Navigation.js
@@ -4,16 +4,16 @@ import { connect } from 'react-redux';
 import { NavLink, withRouter } from 'react-router-dom';
 import Logo from 'component/common/Logo';
 import { confirmLogout } from 'store/actions/session';
-import { SLACK_INFO } from 'global/channels/slack/SlackModels';
-import { EMAIL_INFO } from 'global/channels/email/EmailModels';
-import { JIRA_CLOUD_INFO } from 'global/channels/jira/cloud/JiraCloudModel';
-import { JIRA_SERVER_INFO } from 'global/channels/jira/server/JiraServerModel';
-import { MSTEAMS_INFO } from 'global/channels/msteams/MSTeamsModel';
-import { AZURE_INFO } from 'global/channels/azure/AzureModel';
+import { SLACK_INFO } from 'channels/slack/SlackModels';
+import { EMAIL_INFO } from 'channels/email/EmailModels';
+import { JIRA_CLOUD_INFO } from 'channels/jira/cloud/JiraCloudModel';
+import { JIRA_SERVER_INFO } from 'channels/jira/server/JiraServerModel';
+import { MSTEAMS_INFO } from 'channels/msteams/MSTeamsModel';
+import { AZURE_INFO } from 'channels/azure/AzureModel';
 import { SCHEDULING_INFO } from 'global/components/scheduling/SchedulingModel';
 import { SETTINGS_INFO } from 'global/components/settings/SettingsModel';
 import { AUTHENTICATION_INFO } from 'global/components/auth/AuthenticationModel';
-import { BLACKDUCK_INFO } from 'global/providers/blackduck/BlackDuckModel';
+import { BLACKDUCK_INFO } from 'providers/blackduck/BlackDuckModel';
 import { AUDIT_INFO } from 'global/components/audit/AuditModel';
 import { CERTIFICATE_INFO } from 'global/components/certificates/CertificateModel';
 import { TASK_MANAGEMENT_INFO } from 'global/components/task/TaskManagementModel';
