+++ ui/src/main/js/channels/azure/AzureDistributionConfiguration.js
