+++ ui/src/main/js/channels/azure/AzureGlobalConfiguration.js
@@ -5,7 +5,7 @@ import { CONTEXT_TYPE } from 'util/descriptorUtilities';
 import CommonGlobalConfiguration from 'global/CommonGlobalConfiguration';
 import CommonGlobalConfigurationForm from 'global/CommonGlobalConfigurationForm';
 import PasswordInput from 'field/input/PasswordInput';
-import { AZURE_GLOBAL_FIELD_KEYS, AZURE_INFO } from 'global/channels/azure/AzureModel';
+import { AZURE_GLOBAL_FIELD_KEYS, AZURE_INFO } from 'channels/azure/AzureModel';
 import OAuthEndpointButtonField from 'field/input/OAuthEndpointButtonField';
 import * as GlobalRequestHelper from 'global/GlobalRequestHelper';
 import TextInput from 'field/input/TextInput';
