+++ ui/src/main/js/channels/azure/AzureModel.js
@@ -1,3 +1,16 @@
+export const AZURE_INFO = {
+    key: 'channel_azure_boards',
+    url: 'azure_boards',
+    label: 'Azure Boards'
+};
+
+export const AZURE_GLOBAL_FIELD_KEYS = {
+    organization: 'azure.boards.organization.name',
+    clientId: 'azure.boards.client.id',
+    clientSecret: 'azure.boards.client.secret',
+    configureOAuth: 'azure.boards.oauth'
+};
+
 export const AZURE_DISTRIBUTION_FIELD_KEYS = {
     comment: 'channel.azure.boards.work.item.comment',
     project: 'channel.azure.boards.project',
