+++ ui/src/main/js/channels/email/EmailDistributionConfiguration.js
