+++ ui/src/main/js/channels/jira/server/JiraServerDistributionConfiguration.js
