+++ ui/src/main/js/channels/jira/server/JiraServerModel.js
@@ -1,3 +1,17 @@
+export const JIRA_SERVER_INFO = {
+    key: 'channel_jira_server',
+    url: 'jira_server',
+    label: 'Jira Server'
+};
+
+export const JIRA_SERVER_GLOBAL_FIELD_KEYS = {
+    url: 'jira.server.url',
+    username: 'jira.server.username',
+    password: 'jira.server.password',
+    disablePluginCheck: 'jira.server.disable.plugin.check',
+    configurePlugin: 'jira.server.configure.plugin'
+};
+
 export const JIRA_SERVER_DISTRIBUTION_FIELD_KEYS = {
     comment: 'channel.jira.server.add.comments',
     fieldMapping: 'channel.jira.server.field.mapping',
