+++ ui/src/main/js/channels/msteams/MSTeamsModel.js
@@ -3,3 +3,7 @@ export const MSTEAMS_INFO = {
     url: 'msteams',
     label: 'MS Teams'
 };
+
+export const MSTEAMS_DISTRIBUTION_FIELD_KEYS = {
+    webhook: 'channel.msteams.webhook'
+};
