+++ ui/src/main/js/channels/msteams/MsTeamsDistributionConfiguration.js
