+++ ui/src/main/js/channels/slack/SlackDistributionConfiguration.js
