+++ ui/src/main/js/channels/slack/SlackModels.js
@@ -1,3 +1,9 @@
+export const SLACK_INFO = {
+    key: 'channel_slack',
+    url: 'slack',
+    label: 'Slack'
+};
+
 export const SLACK_DISTRIBUTION_FIELD_KEYS = {
     channelName: 'channel.slack.channel.name',
     channelUser: 'channel.slack.channel.username',
