--- ui/src/main/js/distribution/channels/msteams/MsTeamsModel.js
@@ -1,3 +1,0 @@
-export const MSTEAMS_DISTRIBUTION_FIELD_KEYS = {
-    webhook: 'channel.msteams.webhook'
-};
