--- ui/src/main/js/global/channels/azure/AzureModel.js
@@ -1,12 +1,0 @@
-export const AZURE_INFO = {
-    key: 'channel_azure_boards',
-    url: 'azure_boards',
-    label: 'Azure Boards'
-};
-
-export const AZURE_GLOBAL_FIELD_KEYS = {
-    organization: 'azure.boards.organization.name',
-    clientId: 'azure.boards.client.id',
-    clientSecret: 'azure.boards.client.secret',
-    configureOAuth: 'azure.boards.oauth'
-};
