--- ui/src/main/js/global/channels/jira/cloud/JiraCloudModel.js
@@ -1,13 +1,0 @@
-export const JIRA_CLOUD_INFO = {
-    key: 'channel_jira_cloud',
-    url: 'jira',
-    label: 'Jira Cloud'
-};
-
-export const JIRA_CLOUD_GLOBAL_FIELD_KEYS = {
-    url: 'jira.cloud.url',
-    emailAddress: 'jira.cloud.admin.email.address',
-    accessToken: 'jira.cloud.admin.api.token',
-    disablePluginCheck: 'jira.cloud.disable.plugin.check',
-    configurePlugin: 'jira.cloud.configure.plugin'
-};
