--- ui/src/main/js/global/channels/jira/server/JiraServerModel.js
@@ -1,13 +1,0 @@
-export const JIRA_SERVER_INFO = {
-    key: 'channel_jira_server',
-    url: 'jira_server',
-    label: 'Jira Server'
-};
-
-export const JIRA_SERVER_GLOBAL_FIELD_KEYS = {
-    url: 'jira.server.url',
-    username: 'jira.server.username',
-    password: 'jira.server.password',
-    disablePluginCheck: 'jira.server.disable.plugin.check',
-    configurePlugin: 'jira.server.configure.plugin'
-};
