--- ui/src/main/js/global/channels/slack/SlackModels.js
@@ -1,5 +1,0 @@
-export const SLACK_INFO = {
-    key: 'channel_slack',
-    url: 'slack',
-    label: 'Slack'
-};
