+++ ui/src/main/js/providers/blackduck/BlackDuckModel.js
